Adding the remade shell, with the README and the Man
This commit contain all the main files.
main_controller contians all of the main loop function, that allow you to type in it .
In the str_function files are all the str controls.
in the Path/Tokens/controls has all the function related to finding path, making tokens and controls those tokens